keep y as native dtype

+++ src/quant_analysis/metric_calculation/evaluate_mean_absolute_error.py
@@ -16,7 +16,7 @@ def evaluate_mae(
     with torch.no_grad():
         for batch in dataloader:
             X = batch[0].to(device=device, dtype=input_dtype)
-            y = batch[1].to(device, dtype=input_dtype)
+            y = batch[1].to(device)
             pred = model(X)
             absolute_error += loss_function(pred, y).item()
